webadmin: New disk dialog does not validate the size value
The size field is defined as as an entity model of type long and so the
integer validator does not validate it at all making it possible for numbers
less than the minimum (including negative number and zero) as well as
numbers higher than the maximum to appear as validated, this was fixed
by replacing the entity model for size from type long to integer

Bug-Url: https://bugzilla.redhat.com/1115471

+++ frontend/webadmin/modules/uicommonweb/src/main/java/org/ovirt/engine/ui/uicommonweb/models/vms/DiskModel.java
@@ -17,7 +17,7 @@ public class DiskModel extends Model
     private Disk disk;
     private VM vm;
 
-    private EntityModel<Long> size;
+    private EntityModel<Integer> size;
     private EntityModel<String> alias;
     private EntityModel<String> description;
     private EntityModel<String> sourceStorageDomainName;
@@ -47,11 +47,11 @@ public class DiskModel extends Model
         this.vm = vm;
     }
 
-    public EntityModel<Long> getSize() {
+    public EntityModel<Integer> getSize() {
         return size;
     }
 
-    public void setSize(EntityModel<Long> size) {
+    public void setSize(EntityModel<Integer> size) {
         this.size = size;
     }
 
@@ -136,7 +136,7 @@ public class DiskModel extends Model
     }
 
     public DiskModel() {
-        setSize(new EntityModel<Long>());
+        setSize(new EntityModel<Integer>());
         setAlias(new EntityModel<String>());
         setDescription(new EntityModel<String>());
         setSourceStorageDomainName(new EntityModel<String>());
